backend/requirements.txt
fastapi[standard]
pydantic
pydantic[email]
mysql-connector-python
colorama 
python-dotenv 
PyJWT